requirements.txt
Pillow==8.0.1
PyQt5==5.15.2
